version 열거형 관련 함수 업데이트
주 버전 및 부 버전을 저장하는 타입을 long long에서 int로 변경하였습니다.

+++ include/Aheuiplusplus/version.hpp
@@ -12,10 +12,10 @@ namespace app
 		latest = v1_1,
 	};
 
-	long long get_major(app::version version) noexcept;
-	long long get_minor(app::version version) noexcept;
-	version get_version(long long major) noexcept;
-	version get_version(long long major, long long minor) noexcept;
+	int get_major(app::version version) noexcept;
+	int get_minor(app::version version) noexcept;
+	version get_version(int major) noexcept;
+	version get_version(int major, int minor) noexcept;
 }
 
 #endif
